Refactor draw_cards
