+++ api/shared/game.py
@@ -1,3 +1,6 @@
+from random import sample
+from itertools import islice, product
+
 def get_player_by_nickname(players, nickname):
     filtered_players = [p for p in players if p["nickname"] == nickname]
     if filtered_players and filtered_players[0]:
@@ -42,3 +45,14 @@ def censor_game(game, current_round, player_authenticated, player_nickname):
         "history": game["history"],
         "last_modified": game["last_modified"]
     }
+
+def draw_cards(players):
+    possible_cards = product(range(6), range(4))
+    all_cards_raw = sample(list(possible_cards), sum(int(p["n_cards"]) for p in players))
+    all_cards = [{"value": tup[0], "colour": tup[1]} for tup in all_cards_raw]
+    card_iterator = iter(all_cards)
+    hands = []
+    for player in players:
+        player_hand = list(islice(card_iterator, 0, int(player["n_cards"])))
+        hands.append({"nickname": player['nickname'], "hand": player_hand})
+    return hands
