+++ api/start_game.py
@@ -11,18 +11,7 @@ from shared.response import *
 from shared.db import *
 from shared.api_gateway import parse_event
 from shared.inputs import is_valid_uuid
-
-
-def draw_cards(players):
-    possible_cards = product(range(6), range(4))
-    all_cards_raw = sample(list(possible_cards), sum(int(p["n_cards"]) for p in players))
-    all_cards = [{"value": tup[0], "colour": tup[1]} for tup in all_cards_raw]
-    card_iterator = iter(all_cards)
-    hands = []
-    for player in players:
-        player_hand = list(islice(card_iterator, 0, int(player["n_cards"])))
-        hands.append({"nickname": player['nickname'], "hand": player_hand})
-    return hands
+from shared.game import draw_cards
 
 
 def arrange_players(players):
